// demos/convex/html/convex/_generated/dataModel.d.ts
// Generated by @neovex/codegen. Do not edit by hand.
import type { GenericId } from "convex/values";

type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export type TableNames = "messages";

export type Id<TableName extends string> = GenericId<TableName>;

type DocumentByTable = {
  "messages": {
    _id: Id<"messages">;
    _creationTime: number;
    "author": string;
    "body": string;
  };
};

type IndexNamesByTable = {
  "messages": "by_author";
};

export type Doc<TableName extends TableNames> = DocumentByTable[TableName];

export type DataModel = {
  [TableName in TableNames]: {
    document: DocumentByTable[TableName];
    fieldPaths: keyof DocumentByTable[TableName];
    indexes: IndexNamesByTable[TableName];
  };
};